Improved graphing
Changed the sketch such that min/max of each actuator is printed, followed by filtered and raw position data of each actuator. These are printed with commas separating them and can be easily imported into matlab for processing/graphing. A matlab program for doing so is also in this commit, although changes can be made to it for more/different analysis.